Look Ma!  No Ids on the domain objects!
Added a filter (RestFilter.java) that auto-reroutes dynamic URIs with
embedded variables in the path.

Changed the UrlPatterns in the Servlet Controllers to use the embedded
path variables instead of relying on get/post parameters to provide the
context.

Scrapped storing the DB ids on the domain objects and replaced with
using the natural key when generating the hashCode and using that for
the external identifiers for both the ReST and DB references.

Reworked everything else so that it works again with those changes.